safari-600.7-branch/LayoutTests/editing/inserting/insert-br-quoted-007-expected.txt

CONSOLE MESSAGE: line 18: SUCCESS! newline inserted AFTER quoted content

Line of text
